+++ tests/helpers/load-content-module.js
@@ -19,6 +19,7 @@ function clearContentGlobals() {
     delete globalThis.NSM_CREATE_CONTENT_IMPORT_EXPORT;
     delete globalThis.NSM_CREATE_CONTENT_DIAGNOSTICS;
     delete globalThis.NSM_CREATE_CONTENT_SOURCE_VIEW_SWITCH_CONTROLLER;
+    delete globalThis.NSM_CREATE_CONTENT_SNAPSHOT_SIGNATURE;
     delete globalThis.NSM_CREATE_CONTENT_PERSISTENCE;
     delete globalThis.NSM_CREATE_CONTENT_SOURCE_LIST_SCAN;
     delete globalThis.NSM_CREATE_CONTENT_NATIVE_LABEL_SCAN;
@@ -62,6 +63,7 @@ function loadContentModule() {
     require('../../src/content/content-import-export.js');
     require('../../src/content/content-diagnostics.js');
     require('../../src/content/content-source-view-switch-controller.js');
+    require('../../src/content/content-snapshot-signature.js');
     require('../../src/content/content-persistence.js');
     require('../../src/content/content-source-list-scan.js');
     require('../../src/content/content-native-label-scan.js');